requirements.txt
# Core dependencies
click>=8.1.0
pyyaml>=6.0

# Database
sqlalchemy>=2.0.0

# Cloud provider SDKs (optional)
boto3>=1.28.0  # AWS
azure-identity>=1.14.0  # Azure
azure-mgmt-resource>=23.0.0
azure-mgmt-storage>=21.0.0
azure-mgmt-sql>=4.0.0
azure-mgmt-compute>=30.0.0
google-cloud-storage>=2.10.0  # GCP
google-cloud-compute>=1.14.0

# LLM providers (optional)
anthropic>=0.7.0  # Claude
openai>=1.3.0  # OpenAI GPT

# Machine Learning
scikit-learn>=1.3.0
numpy>=1.24.0
joblib>=1.3.0

# Web framework (for dashboard)
flask>=3.0.0
flask-socketio>=5.3.0
plotly>=5.17.0

# Notifications
requests>=2.31.0

# Utilities
python-dateutil>=2.8.2
